reconfigured folder structure
