+++ app/main.py
@@ -1,0 +1,8 @@
+from fastapi import FastAPI
+from app.routes.user_routes import router as user_router
+app=FastAPI()
+
+app.include_router(user_router)
+
+
+
