+++ app/models/schema.py
@@ -1,9 +1,7 @@
 from pydantic import BaseModel
-from typing import List
 
 class User(BaseModel):
     id:int
     name:str
     address:str
 
-users:List[User]=[]
