+++ app/utils/db.py
@@ -1,0 +1,5 @@
+
+from typing import List
+from app.models.schema import User
+
+users:List[User]=[]
