--- routes/create_user.py
@@ -1,11 +1,0 @@
-from fastapi import APIRouter
-from models import users,User
-
-router=APIRouter()
-
-@router.post("/users")
-def create_user(user:User):
-    users.append(user)
-    return {"message":"New user added successfully"}
-
-
