+++ scripts/start.sh
@@ -1,0 +1,1 @@
+uvicorn app.main:app --reload
