+++ scripts/venv-setup.sh
@@ -1,0 +1,3 @@
+python3 -m venv venv
+source venv/bin/activate
+pip3 install -r requirements.txt
